feat: 🎸 create readTop10ByNumOfReview function

See merge request s07-bigdata-recom-sub2/S07P22B206!54

+++ yogidice/src/main/java/com/specialization/yogidice/domain/repository/category/MechanismRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface MechanismRepository extends JpaRepository<Mechanism, Byte> {
+public interface MechanismRepository extends JpaRepository<Mechanism, Long> {
     Optional<Mechanism> findByName(String name);
 }
