+++ yogidice/src/main/java/com/specialization/yogidice/domain/repository/category/TypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TypeRepository extends JpaRepository<Type, Long> {
+public interface TypeRepository extends JpaRepository<Type, Byte> {
     Optional<Type> findByName(String name);
 }
