+++ yogidice/src/main/java/com/specialization/yogidice/dto/response/category/TypeGroupResponse.java
@@ -12,13 +12,13 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @ApiModel("TypeGroupResponse")
 public class TypeGroupResponse extends BaseResponseBody {
-    private long id;
+    private Long id;
 
     private Long gameId;
 
     private String gameTitle;
 
-    private Long typeId;
+    private Byte typeId;
 
     private String typeName;
